% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparar_Motifs.R
\name{comparar_Motifs}
\alias{comparar_Motifs}
\title{Compara los motif (lista) entre ellos}
\usage{
comparar_Motifs(motif)
}
\arguments{
\item{motif}{Lista de motif}
}
\value{
Matriz de puntuación del alineamiento de los motifs
}
\description{
Función que nos ayuda a observar cuáles son los motifs que se
encuentran más próximos entre ellos a nivel de secuencia.
}
\note{
comparar_Motifs(motif)
}
