% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{compare_Motifs}
\alias{compare_Motifs}
\title{Compare motifs (list)}
\usage{
compare_Motifs(motif, type, method)
}
\arguments{
\item{motif}{Motif in formar list}

\item{type}{The type of matrix that represents the motif and that will be
used for alignment. You can chose between "PPM" or "ICM"}

\item{method}{Alignment metric between motifs. You can chose one method: PCC,
EUCL,SW,KL,ALLR,BHAT,HELL,SEUCL,MAN,ALLR_LL, WEUCL, WPCC. See details}
}
\value{
Motif alignment scoring matrix
}
\description{
Function that compares all motifs to all other motifs and
helps us to observe which are the motifs that are are closer to each other at the sequence level.
}
\note{
compare_Motifs(motif, "type", "method") Dont't forget use the quotes in type and method
}
