% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertirDF_Motif.R
\name{convertirDF_Motif}
\alias{convertirDF_Motif}
\title{Convierte un motif (lista) en un Data Frame}
\usage{
convertirDF_Motif(motif)
}
\arguments{
\item{motif}{Motif que queremos ver}
}
\value{
Motif en formato data.frame
}
\description{
Función que nos ayuda a convertir un motif en formato lista a
formato data.frame. Facilitando su visibilidad y manejo en posteriores
funciones
}
\note{
convertirDF_Motif(motif)

 motifDF <- convertirDF_Motif(motif)
}
