% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertirDF_Motifs.R
\name{convertirDF_Motifs}
\alias{convertirDF_Motifs}
\title{Convierte un motif (lista) en un Data Frame}
\usage{
convertirDF_Motifs(motif)
}
\arguments{
\item{motif}{Motif que queremos ver}
}
\value{
Motif en formato data.frame
}
\description{
Función que nos ayuda a convertir un motif en formato lista a
formato data.frame. Facilitando su visibilidad y manejo en posteriores
funciones
}
\note{
convertirDF_Motifs(motif)

 motifDF <- convertirDF_Motifs(motif)
}
