% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{convertToDNAStringSet}
\alias{convertToDNAStringSet}
\title{Convert a motif list in format ".fasta" to DNAStringSet}
\usage{
convertToDNAStringSet(ourpath, nFlist)
}
\arguments{
\item{motif}{Motif we want to see}
}
\value{
Motif in format data.frame
}
\description{
Function that helps us convert ...
}
\note{
listToDF_Motifs(motif)
}
