% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{create_Motif}
\alias{create_Motif}
\title{Create motifs}
\usage{
create_Motif(consensus, name)
}
\arguments{
\item{consensus}{The consensus of the motif}

\item{name}{Name of motif}

\item{pseudoc}{}
}
\value{
A new motif (class universalmotif)
}
\description{
Function that ...
}
\note{
alphabet default DNA, type default PPM
}
