% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{create_Motifeval}
\alias{create_Motifeval}
\title{Create motif ...}
\usage{
create_Motifeval(evall)
}
\arguments{
\item{evall}{The motif ...}
}
\value{
Add new param  evall
}
\description{
Function that ...
}
\note{
alphabet default DNA, type default PPM
}
