% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{create_Motifpval}
\alias{create_Motifpval}
\title{Create motif ...}
\usage{
create_Motifpval(pval)
}
\arguments{
\item{pval}{The motif ...}
}
\value{
Add new param  pval
}
\description{
Function that ...
}
\note{
alphabet default DNA, type default PPM
}
