% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{create_MotifRefined}
\alias{create_MotifRefined}
\title{Create motifs}
\usage{
create_MotifRefined(scan_target, name)
}
\arguments{
\item{name}{Name of motif.}

\item{motifs}{The list of motif}
}
\value{
A new motif (class universalmotif)
}
\description{
Function that ...
}
\note{
scan50refined <- scan_Motif (motif,seq, th)
50refined <- create_Motif(scan50refined, "motifp30_50refined")
}
