% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{enrich_Motifs}
\alias{enrich_Motifs}
\title{Enrich motifs}
\usage{
enrich_Motifs(motif, seq)
}
\arguments{
\item{seq}{The background sequence}

\item{motifs}{The motif to analyze}
}
\value{
The enrichment analysis of this motif
}
\description{
Function that helps us to make an enrichment analysis
of the refined motif compared to its background sequence.
}
\note{
enrich_Motifs=function(cdk2_100,seq_file_cdk2_100)
}
