% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{etapa5_Motifs}
\alias{etapa5_Motifs}
\title{Etapa 5}
\usage{
etapa5_Motifs(motifcl, seq_file, seqbkg, name)
}
\arguments{
\item{motifcl}{The motif obteined in the cluster}

\item{seq_file}{The target sequences}

\item{seqbkg}{The background sequences}

\item{name}{Name of the refined motif}
}
\value{
The refined motif
}
\description{
Function that ...
}
\note{
example<-etapa5_Motifs(clusterp30[1],seqp30_50,SeqBackground,"Refined_p50")
TODO: Añadir extra info y ¿bkg?
}
