% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{export_Jaspar}
\alias{export_Jaspar}
\title{Export motifs to jaspar file}
\usage{
export_Jaspar(motif, filename)
}
\arguments{
\item{motif}{Motif already merged}

\item{filename}{Name of the file}
}
\value{
The definitive motif
}
\description{
Function that we use to export the motif in Jaspar format
}
\note{
export_Jaspar(motifxxx,"file.txt") *don´t forget the ""
}
