% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{export_Motif}
\alias{export_Motif}
\title{Export motifs to universalmotif file}
\usage{
export_Motif(motif, filename)
}
\arguments{
\item{motif}{Motif already merged}

\item{filename}{Name of the file}
}
\value{
The definitive motif
}
\description{
Function we use to export the motif in universalmotif format
}
\note{
export_Jaspar(motifxxx,"file.txt") *don´t forget the ""
}
