% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{heatMap_Motifs}
\alias{heatMap_Motifs}
\title{Get a heatmap of the selected motifs (list)}
\usage{
heatMap_Motifs(motif)
}
\arguments{
\item{motif}{List of motifs}
}
\value{
Motif alignment scoring matrix
}
\description{
Function that by means of a graphical plot allows us to visualize the
}
\note{
heatMap_Motifs(motif)
}
