% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_Motif.R
\name{leer_Motif}
\alias{leer_Motif}
\title{Importa motifs en formato HOMER}
\usage{
leer_Motif(rutadelmotiv)
}
\arguments{
\item{rutadelmotiv}{Ruta en nuestro pc del motif que queremos importar}
}
\value{
El motif importado con los parametros más destacables
}
\description{
Función que carga y hace comprensible para el humano un motif en formato HOMER
}
\note{
leer_motif(rutadelmotiv = "./ejemplo/ejemplo.motif")
}
