% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leerMotif.R
\name{leerMotif}
\alias{leerMotif}
\title{Importa motifs en formato HOMER}
\usage{
leerMotif(rutadelmotiv)
}
\arguments{
\item{rutadelmotiv}{Ruta en nuestro pc del motif que queremos importar}
}
\value{
El motif importado con los parametros más destacables
}
\description{
Función que carga y hace comprensible para el humano un motif en formato HOMER
}
