% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{listToDF_Motifs}
\alias{listToDF_Motifs}
\title{Convert motif (list) in a Data Frame}
\usage{
listToDF_Motifs(motif)
}
\arguments{
\item{motif}{Motif we want to see}
}
\value{
Motif in format data.frame
}
\description{
Function that helps us convert a motif in list format to
}
\note{
listToDF_Motifs(motif)

motifDF <- listToDF_Motifs(motif)
}
