% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapaCalor_Motifs.R
\name{mapaCalor_Motifs}
\alias{mapaCalor_Motifs}
\title{Obtiene un mapa de calor de los motifs seleccionados (lista)}
\usage{
mapaCalor_Motifs(motif)
}
\arguments{
\item{motif}{Lista de motif}
}
\value{
Matriz de puntuación del alineamiento de los motifs
}
\description{
Función que mediante un plot gráfico nos permite visualizar la
matriz obtenida en la comparacion de motifs, y que nos ayuda a seleccionar el
cluster que vamos buscando.
}
\note{
comparar_Motifs(motif)
}
