% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{merge_Motifs}
\alias{merge_Motifs}
\title{Merge motifs}
\usage{
merge_Motifs(motifs)
}
\arguments{
\item{motifs}{List of motifs we want to merge}
}
\value{
The definitive motif
}
\description{
Function that helps us to merge all the motifs into one,
to obtain the final motif.
}
\note{
merge.motifs <- merge_Motifs=function(motif)
}
