% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puntuacion_Motif.R
\name{puntuacion_Motif}
\alias{puntuacion_Motif}
\title{Puntuación del motif}
\usage{
puntuacion_Motif(motif, x)
}
\arguments{
\item{motif}{El motiv del que se quiere obtener las puntuaciones
#' @param x Numero del motif dentro de la secuencia}
}
\value{
El resultado de la puntuación para ese motif
}
\description{
Función que obtiene la puntuacióin de un motif especificado por
parametro según la formula matematica de (Almagro-Hernández and
Fernández-Breis, 2020)
}
\note{
puntuacion_Motif(motif)
}
