% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puntuacionSeq_Motifs.R
\name{puntuacionSeq_Motifs}
\alias{puntuacionSeq_Motifs}
\title{Puntuación de todos los motif de una secuencia}
\usage{
puntuacionSeq_Motifs(motif)
}
\arguments{
\item{motif}{El motiv del que se quiere obtener las puntuaciones}
}
\value{
El resultado de la puntuación para ese/os motif/s
}
\description{
Función que obtiene la puntuacióin de todos los motifs de una
secuencia pasada como data.frame según la formula matematica de
(Almagro-Hernández and Fernández-Breis, 2020)
}
\note{
m <- leer_motif(rutadelmotiv = "./ejemplo/ejemplo.motif")

motifDF <- convertirDF_Motifs(m)

puntuacionSeq_Motifs(motifDF)
}
