% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{read_Motifs}
\alias{read_Motifs}
\title{Import motifs in format HOMER}
\usage{
read_Motifs(rutadelmotiv)
}
\arguments{
\item{rutadelmotiv}{Path in our pc of the motif that we want to import}
}
\value{
The motif list with the parameters more importants
}
\description{
Function that loading y makes a motif understandable to humans
}
\note{
read_Motifs(rutadelmotiv = "./example/example.motif")
}
