% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{scan_Motif}
\alias{scan_Motif}
\title{Scan sequences for matches to input motifs}
\usage{
scan_Motif(motif, seq, th)
}
\arguments{
\item{motif}{The motif in correct format}

\item{seq}{Sequence obtained in step 1 of each experiment
in DNAStringSet format}

\item{th}{Threshold that determines whether or not a sequence contains the
motif}
}
\value{
The analysis of scanned sequences
}
\description{
This function is used to scan the motif of each set
in its respective set of peaks sequences, in order to obtain the subsequences
of the same length as the motif, that exceed the established threshold.
If that score exceeds the predetermined threshold, it is considered that
there is a match and that subsequence is one of those that the function
returns and will be used to generate the refined motif.
}
\note{
scan_Motif(motif,seq_file,0,70)
}
