% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{score_Motif}
\alias{score_Motif}
\title{Score of motif}
\usage{
score_Motif(motif, x)
}
\arguments{
\item{motif}{The motif in data.frame format from which you want to obtain the
scores}

\item{x}{Motif number within the sequence}
}
\value{
The score result for that motif
}
\description{
Function that gets the punctuation of a motif specified by
parameter according to the mathematical formula of (Almagro-Hernández and
Fernández-Breis, 2020)
}
\note{
score_Motif(motifdf, x)
}
