% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{scoreSeq_Motifs}
\alias{scoreSeq_Motifs}
\title{Score of all motifs in a sequence (data.frame)}
\usage{
scoreSeq_Motifs(motif)
}
\arguments{
\item{motif}{The motif from which you want to obtain the scores}
}
\value{
The score result for that motif / s
}
\description{
Function that obtains the punctuation of all the motifs of a
sequence passed as data.frame according to the mathematical formula of
(Almagro-Hernández and Fernández-Breis, 2020)
}
\note{
m <- read_Motifs(rutadelmotiv = "./ejemplo/ejemplo.motif")

motifDF <- listToDF_Motifs(m)

scoreSeq_Motifs_Motifs(motifDF)
}
