% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{see_Motif}
\alias{see_Motif}
\title{Draw a plot with the indicated motif}
\usage{
see_Motif(motif)
}
\arguments{
\item{motif}{Motif we want to see}
}
\value{
The graphic with the motif sequence logo
}
\description{
Function that draws a graphic of the motif that is indicated as
a parameter
}
\note{
motif <- read_Motifs("./xxx/xxx.motif")

see_Motif(motif)
}
