% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{stage1_Motifs}
\alias{stage1_Motifs}
\title{Stage 1}
\usage{
stage1_Motifs(species, chromosome, start, end)
}
\arguments{
\item{species}{The species name/alias. (EX: human, homo_sapiens, mouse)}

\item{chromosome}{The number of chromosome.}

\item{start}{The start of query region. A maximum of 10Mb is allowed to be
requested at any one time}

\item{end}{The end of query region.}
}
\value{
The sequence in format fasta
}
\description{
Function that returns the genomic sequence of the specified
region of the given species.
}
\note{
The value of the species argument must be loaded.
human <- "human", mouse="mouse" (you need load this value to
make the function work)
stage1_Motifsetapa1_Motifs(human,16185,16235)
}
