% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verMotif.R
\name{verMotif}
\alias{verMotif}
\title{Dibuja un plot con el motif indicado}
\usage{
verMotif(x)
}
\arguments{
\item{x}{Motif que queremos ver}
}
\value{
el gráfico con el logo de secuencia del motif
}
\description{
Función que dibuja un gráfico del motif que se le índica como parametro
}
