% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{convertToDNAStringSet}
\alias{convertToDNAStringSet}
\title{Convert a motif list in format ".fasta" to DNAStringSet}
\usage{
convertToDNAStringSet(ourpath, nFlist)
}
\arguments{
\item{ourpath}{Path where the sequence will be found}

\item{nFlist}{The file .fasta (number in your directory)}
}
\value{
The DNAStringSet file
}
\description{
Function that helps us to convert a background sequence
in a fasta format to one in DNAStringSet format
}
\note{
region <- "peaks" (you need load this value to make the function work)
file<-convertToDNAStringSet=function("ourpath", nFlist)
Example: seq1<-convertToDNAStringSet(".", 1)
seq1<-convertToDNAStringSet(".", 2)
}
