% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{enrich_Motifs}
\alias{enrich_Motifs}
\title{Enrich motif ... TODO}
\usage{
enrich_Motifs(motif, seq)
}
\arguments{
\item{seq}{Sequence we use .. . TODO}

\item{motifs}{Motifs we use .. . TODO}

\item{qval}{Sequence we use .. . TODO}
}
\value{
The motif TODO
}
\description{
Function that ... TODO
}
\note{
enrich...
}
