% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{etapa1_Motifs}
\alias{etapa1_Motifs}
\title{Etapa 1}
\usage{
etapa1_Motifs()
}
\value{
The refined motif
}
\description{
Function that ...
}
\note{
example<-etapa5_Motifs(clusterp30[1],seqp30_50,SeqBackground,"Refined_p50")
}
