% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{heatMap_Motifs}
\alias{heatMap_Motifs}
\title{Get a heatmap of the selected motifs (list)}
\usage{
heatMap_Motifs(matrix)
}
\arguments{
\item{matrix}{Alignment scoring matrix}
}
\value{
Heatmap of the clouter motifs
}
\description{
Function that by means of a graphical plot allows us
to visualize the matrix obtained in the comparison of motifs, and that
helps us to select the cluster that we are looking for.
}
\note{
heatMap_Motifs(compare_Motifs(motif = xxx,"xxx","xxx"))
}
