% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{merge_Motifs}
\alias{merge_Motifs}
\title{Merge motifs}
\usage{
merge_Motifs(motifs, type, method)
}
\arguments{
\item{motifs}{List of motifs we want to merge}

\item{type}{The type of matrix that represents the motif and that will be
used for alignment. You can chose between "PPM" or "ICM"}

\item{method}{Alignment metric between motifs. You can chose one method: PCC,
EUCL,SW,KL,ALLR,BHAT,HELL,SEUCL,MAN,ALLR_LL, WEUCL, WPCC. See details}
}
\value{
The definitive motif
}
\description{
Function that helps us to merge all the motifs into one,
to obtain the final motif.
}
\note{
merge.motifs <- merge_Motifs=function(motif,"PPM", "PCC")
}
