% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{merge_Motifs}
\alias{merge_Motifs}
\title{Merge motifs ... TODO funciona bien}
\usage{
merge_Motifs(motifs)
}
\arguments{
\item{motifs}{List of motif we want to merge}
}
\value{
The motifs merged
}
\description{
Function that helps ... TODO
}
\note{
merge.motifs <- merge_Motifs=function(motif)
view_motifs(merge_motif(m1, m2, m3))
}
