% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{scan_Motif}
\alias{scan_Motif}
\title{Scan sequences for matches to input motifs}
\usage{
scan_Motif(motif, seq, th)
}
\arguments{
\item{motif}{The motif in data.frame format from which you want to obtain the scores}

\item{seq}{TODO ...}

\item{th}{TODO ...}
}
\value{
TODO ...
}
\description{
TODO ...
}
\note{
score_Motif(motifdf, x)
}
