% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{stage5_Motifs}
\alias{stage5_Motifs}
\title{Stage 5}
\usage{
stage5_Motifs(motifcl, seq_file, seqbkg, name, th)
}
\arguments{
\item{motifcl}{The motif obteined in the cluster}

\item{seq_file}{The target sequences}

\item{seqbkg}{The background sequences}

\item{name}{Name of the refined motif}

\item{th}{Threshold that determines whether or not a sequence contains the motif}
}
\value{
The refined motif
}
\description{
Function that we use to obtain the refined motif
}
\note{
example<-stage5_Motifs(clusterp30[1],seqp30_50,SeqBackground,"Refined_p50",0.7)
TODO: Añadir extra info y ¿bkg?
}
