% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Motifs.R
\name{trim_Motifs}
\alias{trim_Motifs}
\title{Trim motifs}
\usage{
trim_Motifs(motif)
}
\arguments{
\item{motif}{Motif already merged}
}
\value{
The definitive motif
}
\description{
Function that ... TODO
}
\note{
def.motif <- trim_Motifs(motif)
}
