% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ver_Motif.R
\name{ver_Motif}
\alias{ver_Motif}
\title{Dibuja un plot con el motif indicado}
\usage{
ver_Motif(x)
}
\arguments{
\item{x}{Motif que queremos ver}
}
\value{
El gráfico con el logo de secuencia del motif
}
\description{
Función que dibuja un gráfico del motif que se le índica como parametro
}
\note{
motif <- leer_Motif("./xxx/xxx.motif")

ver_Motif(motif[1])
}
